with jpaa and without spring rest

+++ Shopping/src/main/java/com/online/dao/CategoryDao.java
@@ -1,14 +1,16 @@
 package com.online.dao;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.online.model.Book;
 import com.online.model.BookCategory;
 
 @Repository
-public interface CategoryDao extends JpaRepository<BookCategory,Long> {
-
-	
+public interface CategoryDao extends JpaRepository<BookCategory,Integer> {
 	
 
+	
 }
